Runtime IO (#189)
_Based on #173._

This PR experiments with how Input / Output can work. It is based on
MMIO - Memory-Mapped IO. It works by reading or writing to specific
memory addresses. The system recognizes this and can commit to the
content as part of the proof.

This can also be used for debug statements during development:

```
📜📜📜 Hello, World!
🌏🌍🌎
test test_ceno_rt_io ... ok
```

---------

Co-authored-by: Aurélien Nicolas <info@nau.re>