refactor/rename UInt (#240)
A quick PR, just rename the following 3 structues
- UInt --> UIntLimbs
- RegUInt --> UInt
- UIntValue --> Value

`RegUInt` is designed for constraint system and I think `UInt` is enough
to represent it. `UIntValue` is designed for witness assignment and
looks similiar with `RegUInt`. To avoid being confused for new comers,
rename to `Vaule` (follow zkEVM's naming). After `RegUInt` being
renamed, the original `UInt` conflicts with it so renaming the original
`UInt` to `UIntLimbs`.

cc @hero78119 and @kunxian-xia