Explicit trait for register values (#256)
The issue is that different instructions need different register
representations. The values must be connected through a common format.

Currently, that format is quasi invisible when looking at `trait
ToExpr`. In fact, the types `Uint` and `Uint8` both appear to work but
their values are incompatible (`2 * u16` vs `4 * u8`).

In this PR:
- Introduce a trait specifically for the register format.
- Implement both `Uint` and `Uint8` such that they are compatible with
each other.

Besides, after that we have the option of changing this format. In
particular, it should be the common denominator which is just one
expression of `u32`.

---------

Co-authored-by: Aurélien Nicolas <info@nau.re>