CI: pin rust-toolchain to nightly-1.80.0 (#262)
The current rust-toolchain is `nightly` which changes over the time,
therefore it is more vulnerable to produce unexpected behavior in CI
than a "stable" nightly build. We did not use stable release of Rust
because crate `goldilocks` relies on the feature `generic_const_exprs`
that is only available on nightly build.