chore: make sure all text files end with a newline (#312)
The [POSIX standard
says](https://stackoverflow.com/questions/729692/why-should-text-files-end-with-a-newline)
all text files should end with a newline.

Not that the standard matters that much, but git's diff output is a bit
nicer without the `\ No newline at end of file` warning, and some
command line tools work better if all lines terminate with a newline.
Similar for github's diff view.

If you are curious, you can use ripgrep (or a similar tool) to find all
offending files:

```bash
rg --files-with-matches --multiline '[^\n]\z'
```

Co-authored-by: Ming <hero78119@gmail.com>