Fix 'Unchanged files with check annotations' spam (#338)
On every PR GitHub helpfully complains about 'Unchanged files with check
annotations'. So here we silence most of that, to make all our PRs
easier to review. Either by outright fixing the issue, or by flagging it
to the compiler as a known problem.

This should make it easier for me to review eg
https://github.com/scroll-tech/ceno/pull/335 (or any other PR).