Run clippy everywhere (#372)
Instead of whitelisting a few packages to run clippy on, we just
blacklist the two where it's not expected to work.

Based on https://github.com/scroll-tech/ceno/pull/371. So when that PR
merges, this PR's diff will become smaller. But we might still have to
break this one down into multiple PRs.