chore: update toolchain (#309)
Our `rust-toolchain` specifies `nightly-2024-05-02`. That one comes with
`rustfmt` version 1.7. But our `rustfmt.toml` uses `style_edition =
"2024"`; but `2024` was only added in version 1.8. So in this PR we
update to a more recent nightly toolchain.

Most of the changes in here are from our specified `rustfmt.toml` config
actually applying. The main thing that's causing a lot of diff is
[version sorting](https://github.com/rust-lang/rust/issues/123800) of
imports, which changes how letter case is handled for comparison.

Fixes https://github.com/scroll-tech/ceno/issues/297