Feat: expose `base_address` and `instructions` decoded from ELF (#487)
This PR aims to extract more information from ELF file. The program
table circuit needs the whole instructions to build the fixed program
table.

Meanwhile, we also distinguish two values of program counter: 
- `pc_base` (was former called `pc_start` incorrectly): It's address of
the first opcode in the program. This is equal to
`Platform::rom_start()`.
- `entrypoint`(this should be `pc_start`): It's the address of `.start`
label in the assembly source of guest program and address of the first
opcode to be executed.