Recognise that `namespace` can't fail (#490)
At least `ConstraintSystem::namespace` won't fail on its own. But it
might fail, if the circus construction function (or so) you are passing
to fails.

The new type reflects that.

Extracted from https://github.com/scroll-tech/ceno/pull/466 and
addressing
https://github.com/scroll-tech/ceno/pull/455/files#r1818907561

This will lead to a lot of clean up down the line, but I want to keep
this PR small.