Remove `MRET` (#497)
There are many extensions to RiscV. We only support one of them: 'RV32M'
(which includes multiplication and division).

Generally, we ignore the unsupported extenions. There's no trace of
floating point or compressed instructions in our code.

For some reason a snippet of one of the priviliged instructions has made
it into our code: MRET. We don't actually support it anyway, so we might
as well remove it. Just like we have no trace of any of the other
unsupported extensions in our code.