Feat/MULH opcode (#492)
Implement MULH opcode

Circuit is validated by the following strategy:
1. Compute the signed values associated with `rs1`, `rs2` and `rd`
registers
2. Verify that the product of signed inputs `rs1` and `rs2` is equal to
the result of interpreting `rd` as the high limb of a 2s complement
value with some 32-bit low limb.

The latter value in step 2 is expressed as: the 32-bit signed 2s
complement value of the `rd` register times `2^32`, plus the unsigned
value of the 32-bit low limb.

The soundness here is a bit subtle. The signed values of 32-bit inputs
`rs1` and `rs2` have values between `-2^31` and `2^31 - 1`, so their
product is constrained to lie between `-2^62 + 2^31` and `2^62`. In a
prime field of size smaller than `2^64`, the range of values represented
by a 64-bit 2s complement value, integers between `-2^63` and `2^63 -
1`, have some ambiguity. If `p = 2^64 - k`, then the values between
`-2^63` and `-2^63 + k - 1` correspond with the values between `2^63 -
k` and `2^63 - 1`.

However, as long as the values required by signed products don't overlap
with this ambiguous range, an arbitrary 64-bit 2s complement value can
represent a signed 32-bit product in only one way, so there is no
ambiguity in the representation. This is the case for the Goldilocks
field with order `p = 2^64 - 2^32 + 1`.

---------

Co-authored-by: Bryan Gillespie <bryan@inversed.tech>
Co-authored-by: naure <naure@users.noreply.github.com>