Switch to `InsnKind` in the program table (#538)
_Matthias:_

Risc-V instruction are encoded in multiple different 'types'. These
types have various fields. One particularly common field is `opcode`,
which typically (but not always) occupies the 7 least significant bits.

As the name suggests, opcode helps code _which_ operation is intended. I
say 'helps' because some operations share the same opcode.

Our code before this PR assumed that bits `12..=14` in the instruction
word are exactly enough extra information to determine the operation.
(Confusingly enough, our code follows Risc0 convention, and calls that
range of bits `funct3`. Only some operations have a field named `funct3`
in the spec, and not all of them have it in the same place.) That
assumption is wrong, and our circuits are unsound. Specifically, the
distinction between 'signed' (ie 'arithmetic') and 'unsigned' (ie
'logical') immediate right shifts is encoded in bit 30, as laid out in
section 2.4.1 of the
[spec](https://github.com/riscv/riscv-isa-manual/releases/download/20240411/unpriv-isa-asciidoc.pdf).

We could introduce a new column called 'bit 30' in the program table to
fix the symptom. But instead of such a piecemeal approach, we attack the
underlying problem: we switch the program table to record `InsnKind`.
`InsnKind` directly reports the kind of instruction we are dealing with.
We already have that information available in the decoder, and just
re-use it here.

Fixes https://github.com/scroll-tech/ceno/issues/539 and
https://github.com/scroll-tech/ceno/issues/533

---------

Co-authored-by: Aurélien Nicolas <info@nau.re>