Remove `Rng` from PCS type paramter. (#549)
Extracting a small PR from #294.

The random generator is used only when the encoding scheme is Basecode,
and not used for RS code. So remove it from the type parameter for
Basefold, i.e., `Basefold<E, Spec, Rng>` is replaced with `Basefold<E,
Spec>` to make the API simpler. Fix it to `ChaCha8Rng` in the
implementation of Basecode.