Feat: allow constraint_system to take record as input instead of record's rlc (#569)
This PR is extracted from #526. 

In practice, when we use mock prover to find out constraint errors like
r/w mismatch, we want to know the exact value of these records. However,
currently we only store the RLC of record in the `ConstraintSystem`.
Therefore in this PR we change the interface of `ConstraintSystem`.