CI with a slimmer cargo cache (#748)
I suspect `~/.cargo` gets preserved on our self-hosted runner anyway?

In any case, the self-hosted runner is pretty slow at loading the cache
from github. So we make that faster here.

For reviewers: I suggest to compare the timings on the integrations
check to timings in `master`. The idea is to speed that one up.