Update assignment interface for IsLtConfig gadget (#774)
The existing implementation for `InnerLtConfig` allows assignment from
`u64` values and `SWord` values, the latter of which is an alias for
`i32`. For the signed case, it is simple to extend this support to `i64`
inputs because the only way that the input data type influences the
assignment operation is in the output of the operation `lhs < rhs`, and
`i64` inputs transparently represent `i32` comparison operations after
casting. Allowing `i64` assignment will be helpful for the DIV/REM
opcodes circuit, where an edge case relies on being able to express in
particular `2^32` as a signed value for comparison.

This PR additionally adds documentation for the `InnerLtConfig` gadget,
and simplifies the signed and unsigned assignment cases by implementing
a new `assign_instance_field` function call which takes as input two
basefield elements along with an explicit boolean value for the `is_lt`
input of the circuit.

---------

Co-authored-by: Bryan Gillespie <bryan@inversed.tech>