Feat/Opcode Circuits for DIV, REM, and REMU opcodes (#596)
This PR extends the existing `DIVU` opcode implementation to expose the
remainder as an output, and to handle the (more complicated) signed
case. See the comments in `div.rs` for details, but in summary:
- The basic design of signed and unsigned `DIV` and `REM` opcode
circuits is to represent the division algorithm relation `dividend =
quotient * divisor + remainder` using native field elements, and
constraining so that either `0 <= remainder < divisor` or `divisor == 0`
holds.
- Some nontrivial steps are needed in the signed case to handle the
additional complexity of different signs for the divisor and remainder
values, as well as to handle signed division overflow where `-2^31 / -1
= 2^31` is out of the representable range for `i32` values.
- In either case, either the `quotient` or the `remainder` value is
exposed as the output register `rd`.
- In theory with the current implementation, consecutive calls to `DIV`
and `REM` or `DIVU` and `REMU` could be batched into a single circuit
since both of the outputs are represented; however, this is not yet
implemented.

Currently the circuit implementation and documentation are finished and
ready for review. Unit tests still need to be finished, but it would be
great to get some eyes on the circuit design now.

---------

Co-authored-by: Bryan Gillespie <bryan@inversed.tech>
Co-authored-by: Mihai <mihai.calancea@gmail.com>